updated files for SPINODE
